refactor SpacetimeAggregation into spacetime module
