+++ collate/collate.py
@@ -296,139 +296,3 @@ class Aggregation(object):
         conn.execute(self.get_drop())
         conn.execute(self.get_create())
         trans.commit()
-
-
-class SpacetimeAggregation(Aggregation):
-    def __init__(self, aggregates, groups, intervals, from_obj, dates,
-                 prefix=None, suffix=None, schema=None, date_column=None, output_date_column=None):
-        """
-        Args:
-            aggregates: collection of Aggregate objects
-            from_obj: defines the from clause, e.g. the name of the table
-            groups: a list of expressions to group by in the aggregation or a dictionary
-                pairs group: expr pairs where group is the alias (used in column names)
-            intervals: the intervals to aggregate over. either a list of
-                datetime intervals, e.g. ["1 month", "1 year"], or
-                a dictionary of group : intervals pairs where
-                group is a group in groups and intervals is a collection
-                of datetime intervals, e.g. {"address_id": ["1 month", "1 year]}
-            dates: list of PostgreSQL date strings,
-                e.g. ["2012-01-01", "2013-01-01"]
-            prefix: prefix for column names, defaults to from_obj
-            suffix: suffix for aggregation table, defaults to "aggregation"
-            date_column: name of date column in from_obj, defaults to "date"
-            output_date_column: name of date column in aggregated output, defaults to "date"
-
-        The from_obj and group arguments are passed directly to the
-            SQLAlchemy Select object so could be anything supported there.
-            For details see:
-            http://docs.sqlalchemy.org/en/latest/core/selectable.html
-        """
-        Aggregation.__init__(self,
-                             aggregates=aggregates,
-                             from_obj=from_obj,
-                             groups=groups,
-                             prefix=prefix,
-                             suffix=suffix,
-                             schema=schema)
-
-        if isinstance(intervals, dict):
-            self.intervals = intervals
-        else:
-            self.intervals = {g: intervals for g in self.groups}
-        self.dates = dates
-        self.date_column = date_column if date_column else "date"
-        self.output_date_column = output_date_column if output_date_column else "date"
-
-    def _get_aggregates_sql(self, interval, date, group):
-        """
-        Helper for getting aggregates sql
-        Args:
-            interval: SQL time interval string, or "all"
-            date: SQL date string
-            group: group clause, for naming columns
-        Returns: collection of aggregate column SQL strings
-        """
-        if interval != 'all':
-            when = "{date_column} >= '{date}'::date - interval '{interval}'".format(
-                    interval=interval, date=date, date_column=self.date_column)
-        else:
-            when = None
-
-        prefix = "{prefix}_{group}_{interval}_".format(
-                prefix=self.prefix, interval=interval,
-                group=group)
-
-        return chain(*(a.get_columns(when, prefix, format_kwargs={"collate_date": date})
-                       for a in self.aggregates))
-
-    def get_selects(self):
-        """
-        Constructs select queries for this aggregation
-
-        Returns: a dictionary of group : queries pairs where
-            group are the same keys as groups
-            queries is a list of Select queries, one for each date in dates
-        """
-        queries = {}
-
-        for group, groupby in self.groups.items():
-            intervals = self.intervals[group]
-            queries[group] = []
-            for date in self.dates:
-                columns = [groupby,
-                           ex.literal_column("'%s'::date"
-                                             % date).label(self.output_date_column)]
-                columns += list(chain(*(self._get_aggregates_sql(
-                        i, date, group) for i in intervals)))
-
-                # upper bound on date_column by date
-                where = ex.text("{date_column} < '{date}'".format(
-                        date_column=self.date_column, date=date))
-
-                gb_clause = make_sql_clause(groupby, ex.literal_column)
-                query = ex.select(columns=columns, from_obj=self.from_obj)\
-                          .where(where)\
-                          .group_by(gb_clause)
-
-                if 'all' not in intervals:
-                    greatest = "greatest(%s)" % str.join(
-                            ",", ["interval '%s'" % i for i in intervals])
-                    query = query.where(ex.text(
-                        "{date_column} >= '{date}'::date - {greatest}".format(
-                            date_column=self.date_column, date=date,
-                            greatest=greatest)))
-
-                queries[group].append(query)
-
-        return queries
-
-    def get_indexes(self):
-        """
-        Generate create index queries for this aggregation
-
-        Returns: a dictionary of group : index pairs where
-            group are the same keys as groups
-            index is a raw create index query for the corresponding table
-        """
-        return {group: "CREATE INDEX ON %s (%s, %s);" %
-                (self.get_table_name(group), groupby, self.output_date_column)
-                for group, groupby in self.groups.items()}
-
-    def get_create(self, join_table=None):
-        """
-        Generate a single aggregation table creation query by joining
-            together the results of get_creates()
-        Returns: a CREATE TABLE AS query
-        """
-        if not join_table:
-            join_table = '(%s) t1' % self.get_join_table()
-
-        query = ("SELECT * FROM %s\n"
-                 "CROSS JOIN (select unnest('{%s}'::date[]) as %s) t2\n") % (
-                join_table, str.join(',', self.dates), self.output_date_column)
-        for group, groupby in self.groups.items():
-            query += "LEFT JOIN %s USING (%s, %s)" % (
-                    self.get_table_name(group), groupby, self.output_date_column)
-
-        return "CREATE TABLE %s AS (%s);" % (self.get_table_name(), query)
