+++ tests/test_integration.py
@@ -13,7 +13,8 @@ import sqlalchemy
 from os import path
 import sqlalchemy.sql.expression as ex
 
-from collate import collate
+from collate.collate import Aggregation, Aggregate
+from collate.spacetime import SpacetimeAggregation
 
 with open(path.join(path.dirname(__file__), "config/database.yml")) as f:
     config = yaml.load(f)
@@ -23,8 +24,8 @@ def test_engine():
     assert len(engine.execute("SELECT * FROM food_inspections").fetchall()) == 966
 
 def test_st_explicit_execute():
-    agg = collate.Aggregate("results='Fail'",["count"])
-    st = collate.SpacetimeAggregation([agg],
+    agg = Aggregate("results='Fail'",["count"])
+    st = SpacetimeAggregation([agg],
         from_obj = ex.table('food_inspections'),
         groups = {'license':ex.column('license_no'), 
             'zip':ex.column('zip')},
@@ -37,8 +38,8 @@ def test_st_explicit_execute():
     st.execute(engine.connect())
 
 def test_st_lazy_execute():
-    agg = collate.Aggregate("results='Fail'",["count"])
-    st = collate.SpacetimeAggregation([agg],
+    agg = Aggregate("results='Fail'",["count"])
+    st = SpacetimeAggregation([agg],
         from_obj = 'food_inspections',
         groups = ['license_no', 'zip'],
         intervals = {'license_no':["1 year", "2 years", "all"],
@@ -49,8 +50,8 @@ def test_st_lazy_execute():
     st.execute(engine.connect())
 
 def test_st_execute_broadcast_intervals():
-    agg = collate.Aggregate("results='Fail'",["count"])
-    st = collate.SpacetimeAggregation([agg],
+    agg = Aggregate("results='Fail'",["count"])
+    st = SpacetimeAggregation([agg],
         from_obj = 'food_inspections',
         groups = ['license_no', 'zip'],
         intervals = ["1 year", "2 years", "all"],
@@ -60,16 +61,16 @@ def test_st_execute_broadcast_intervals():
     st.execute(engine.connect())
 
 def test_execute():
-    agg = collate.Aggregate("results='Fail'",["count"])
-    st = collate.Aggregation([agg],
+    agg = Aggregate("results='Fail'",["count"])
+    st = Aggregation([agg],
         from_obj = 'food_inspections',
         groups = ['license_no', 'zip'])
 
     st.execute(engine.connect())
 
 def test_execute_schema_output_date_column():
-    agg = collate.Aggregate("results='Fail'",["count"])
-    st = collate.SpacetimeAggregation([agg],
+    agg = Aggregate("results='Fail'",["count"])
+    st = SpacetimeAggregation([agg],
         from_obj = 'food_inspections',
         groups = ['license_no', 'zip'],
         intervals = {'license_no':["1 year", "2 years", "all"],
